Refactor out the FAQ component
