+++ src/pages/camp/faqs.tsx
@@ -1,68 +1,10 @@
 import * as React from "react"
 import { PageProps, graphql } from "gatsby"
-import {
-  Box,
-  Button,
-  Collapse,
-  Divider,
-  List,
-  ListItem,
-  ListItemButton,
-  ListItemText,
-  Stack,
-  Typography,
-} from "@mui/material"
-import { ExpandLess, ExpandMore } from "@mui/icons-material"
+import { Button, List, ListItem, Stack, Typography } from "@mui/material"
 
 import { Section } from "@components/Layout"
-import PortableText from "@components/PortableText"
 import getPageTitle from "@utils/getPageTitle"
-
-type FAQProps = {
-  question: {
-    readonly questionTitle: string | null
-    readonly _rawQuestionAnswer: Record<string, unknown> | null
-  } | null
-}
-
-function FAQ({ question }: FAQProps) {
-  const [open, setOpen] = React.useState(false)
-  if (!question) {
-    console.warn("Question not found")
-    return <></>
-  }
-
-  const { questionTitle, _rawQuestionAnswer } = question
-  const handleClick = () => {
-    setOpen(!open)
-  }
-
-  return (
-    <>
-      <Divider />
-      <ListItemButton
-        sx={{
-          paddingTop: 2,
-          paddingBottom: 2,
-        }}
-        onClick={handleClick}
-      >
-        <ListItemText
-          primary={questionTitle}
-          primaryTypographyProps={{
-            variant: "h6",
-          }}
-        />
-        {open ? <ExpandLess color="primary" /> : <ExpandMore color="primary" />}
-      </ListItemButton>
-      <Collapse in={open} unmountOnExit>
-        <Box sx={{ padding: 4 }}>
-          <PortableText content={_rawQuestionAnswer} />
-        </Box>
-      </Collapse>
-    </>
-  )
-}
+import FAQ from "@components/FAQ"
 
 export const query = graphql`
   query FAQsPage {
